+++ src/objects/alert-object.ts
@@ -2,7 +2,7 @@ import {
   BLUE_TEAM_COLOR,
   RED_TEAM_COLOR,
 } from "../constants/colors-constants.js";
-import { TimerService } from "../services/timer-service.js";
+import { TimerService } from "../services/gameplay/timer-service.js";
 import { BaseAnimatedGameObject } from "./base/base-animated-object.js";
 import type { MultiplayerGameObject } from "../interfaces/objects/multiplayer-game-object.js";
