+++ src/screens/base/base-game-screen.ts
@@ -5,7 +5,7 @@ import type { GameObject } from "../../interfaces/objects/game-object.js";
 import type { GameScreen } from "../../interfaces/screens/game-screen.js";
 import type { ScreenManager } from "../../interfaces/screens/screen-manager.js";
 import { ScreenManagerService } from "../../services/screen-manager-service.js";
-import { EventConsumerService } from "../../services/event-consumer-service.js";
+import { EventConsumerService } from "../../services/gameplay/event-consumer-service.js";
 import { EventType } from "../../enums/event-type.js";
 import type { GameState } from "../../models/game-state.js";
 import { ServiceLocator } from "../../services/service-locator.js";
