+++ src/screens/main-screen/main-menu-screen.ts
@@ -2,10 +2,10 @@ import { CloseableMessageObject } from "../../objects/common/closeable-message-o
 import { MenuOptionObject } from "../../objects/common/menu-option-object.js";
 import { TitleObject } from "../../objects/common/title-object.js";
 import { ServerMessageWindowObject } from "../../objects/server-message-window-object.js";
-import { APIService } from "../../services/api-service.js";
+import { APIService } from "../../services/network/api-service.js";
 import type { MessagesResponse } from "../../interfaces/responses/messages-response.js";
 import { BaseGameScreen } from "../base/base-game-screen.js";
-import { LoadingScreen } from "../loading-screen.js";
+import { LoadingScreen } from "../loading/loading-screen.js";
 import { ScoreboardScreen } from "./scoreboard-screen.js";
 import { SettingsScreen } from "./settings-screen.js";
 import { EventType } from "../../enums/event-type.js";
