+++ src/screens/main-screen/main-screen.ts
@@ -1,8 +1,8 @@
-import { BaseGameScreen } from "./base/base-game-screen.js";
-import { ScreenManagerService } from "../services/screen-manager-service.js";
-import { MainBackgroundObject } from "../objects/backgrounds/main-background-object.js";
-import type { GameScreen } from "../interfaces/screens/game-screen.js";
-import type { GameState } from "../models/game-state.js";
+import { BaseGameScreen } from "../base/base-game-screen.js";
+import { ScreenManagerService } from "../../services/screen-manager-service.js";
+import { MainBackgroundObject } from "../../objects/backgrounds/main-background-object.js";
+import type { GameScreen } from "../../interfaces/screens/game-screen.js";
+import type { GameState } from "../../models/game-state.js";
 
 export class MainScreen extends BaseGameScreen {
   private screen: GameScreen | null = null;
