+++ src/screens/world/world-screen.ts
@@ -1,35 +1,35 @@
-import { LocalCarObject } from "../objects/local-car-object.js";
-import { WorldBackgroundObject } from "../objects/backgrounds/world-background-object.js";
-import { GoalObject } from "../objects/goal-object.js";
-import { BallObject } from "../objects/ball-object.js";
-import { ScoreboardObject } from "../objects/scoreboard-object.js";
-import { BaseCollidingGameScreen } from "./base/base-colliding-game-screen.js";
-import { GameState } from "../models/game-state.js";
-import { getConfigurationKey } from "../utils/configuration-utils.js";
-import { SCOREBOARD_SECONDS_DURATION } from "../constants/configuration-constants.js";
-import { AlertObject } from "../objects/alert-object.js";
-import { ToastObject } from "../objects/common/toast-object.js";
-import { RemoteCarObject } from "../objects/remote-car-object.js";
-import { ObjectStateType } from "../enums/object-state-type.js";
-import { EventType } from "../enums/event-type.js";
-import { RemoteEvent } from "../models/remote-event.js";
-import { ScreenType } from "../enums/screen-type.js";
-import { MatchStateType } from "../enums/match-state-type.js";
-import type { PlayerConnectedPayload } from "../interfaces/events/player-connected-payload.js";
-import type { PlayerDisconnectedPayload } from "../interfaces/events/player-disconnected-payload.js";
-import { BinaryWriter } from "../utils/binary-writer-utils.js";
-import { BinaryReader } from "../utils/binary-reader-utils.js";
-import type { IMatchmakingProvider } from "../interfaces/services/matchmaking-provider.js";
-import { MatchmakingService } from "../services/matchmaking-service.js";
-import { MatchmakingControllerService } from "../services/matchmaking-controller-service.js";
-import { ScoreManagerService } from "../services/score-manager-service.js";
-import { ServiceLocator } from "../services/service-locator.js";
-import { EventProcessorService } from "../services/event-processor-service.js";
-import { ObjectOrchestratorService } from "../services/object-orchestrator-service.js";
-import { ScreenTransitionService } from "../services/screen-transition-service.js";
-import { TimerManagerService } from "../services/timer-manager-service.js";
-import { MainScreen } from "./main-screen.js";
-import { MainMenuScreen } from "./main-screen/main-menu-screen.js";
+import { LocalCarObject } from "../../objects/local-car-object.js";
+import { WorldBackgroundObject } from "../../objects/backgrounds/world-background-object.js";
+import { GoalObject } from "../../objects/goal-object.js";
+import { BallObject } from "../../objects/ball-object.js";
+import { ScoreboardObject } from "../../objects/scoreboard-object.js";
+import { BaseCollidingGameScreen } from "../base/base-colliding-game-screen.js";
+import { GameState } from "../../models/game-state.js";
+import { getConfigurationKey } from "../../utils/configuration-utils.js";
+import { SCOREBOARD_SECONDS_DURATION } from "../../constants/configuration-constants.js";
+import { AlertObject } from "../../objects/alert-object.js";
+import { ToastObject } from "../../objects/common/toast-object.js";
+import { RemoteCarObject } from "../../objects/remote-car-object.js";
+import { ObjectStateType } from "../../enums/object-state-type.js";
+import { EventType } from "../../enums/event-type.js";
+import { RemoteEvent } from "../../models/remote-event.js";
+import { ScreenType } from "../../enums/screen-type.js";
+import { MatchStateType } from "../../enums/match-state-type.js";
+import type { PlayerConnectedPayload } from "../../interfaces/events/player-connected-payload.js";
+import type { PlayerDisconnectedPayload } from "../../interfaces/events/player-disconnected-payload.js";
+import { BinaryWriter } from "../../utils/binary-writer-utils.js";
+import { BinaryReader } from "../../utils/binary-reader-utils.js";
+import type { IMatchmakingProvider } from "../../interfaces/services/matchmaking-provider.js";
+import { MatchmakingService } from "../../services/gameplay/matchmaking-service.js";
+import { MatchmakingControllerService } from "../../services/gameplay/matchmaking-controller-service.js";
+import { ScoreManagerService } from "../../services/gameplay/score-manager-service.js";
+import { ServiceLocator } from "../../services/service-locator.js";
+import { EventProcessorService } from "../../services/gameplay/event-processor-service.js";
+import { ObjectOrchestratorService } from "../../services/gameplay/object-orchestrator-service.js";
+import { ScreenTransitionService } from "../../services/screen-transition-service.js";
+import { TimerManagerService } from "../../services/gameplay/timer-manager-service.js";
+import { MainScreen } from "../main-screen/main-screen.js";
+import { MainMenuScreen } from "../main-screen/main-menu-screen.js";
 
 export class WorldScreen extends BaseCollidingGameScreen {
   private readonly COUNTDOWN_START_NUMBER = 4;
