+++ src/services/gameplay/game-loop-service.ts
@@ -1,27 +1,27 @@
-import { GameFrame } from "../models/game-frame.js";
-import { NotificationObject } from "../objects/common/notification-object.js";
-import { MainScreen } from "../screens/main-screen.js";
-import { LoginScreen } from "../screens/main-screen/login-screen.js";
-import { MainMenuScreen } from "../screens/main-screen/main-menu-screen.js";
-import { EventType } from "../enums/event-type.js";
-import type { ServerDisconnectedPayload } from "../interfaces/events/server-disconnected-payload.js";
-import type { ServerNotificationPayload } from "../interfaces/events/server-notification-payload.js";
-import { CANVAS_HEIGHT, CANVAS_WIDTH } from "../constants/canvas-constants.js";
-import { DebugUtils } from "../debug/debug-utils.js";
-import type { GameScreen } from "../interfaces/screens/game-screen.js";
-import { GAME_VERSION } from "../constants/game-constants.js";
+import { GameFrame } from "../../models/game-frame.js";
+import { NotificationObject } from "../../objects/common/notification-object.js";
+import { MainScreen } from "../../screens/main-screen/main-screen.js";
+import { LoginScreen } from "../../screens/main-screen/login-screen.js";
+import { MainMenuScreen } from "../../screens/main-screen/main-menu-screen.js";
+import { EventType } from "../../enums/event-type.js";
+import type { ServerDisconnectedPayload } from "../../interfaces/events/server-disconnected-payload.js";
+import type { ServerNotificationPayload } from "../../interfaces/events/server-notification-payload.js";
+import { CANVAS_HEIGHT, CANVAS_WIDTH } from "../../constants/canvas-constants.js";
+import { DebugUtils } from "../../debug/debug-utils.js";
+import type { GameScreen } from "../../interfaces/screens/game-screen.js";
+import { GAME_VERSION } from "../../constants/game-constants.js";
 import { EventConsumerService } from "./event-consumer-service.js";
-import { DebugObject } from "../debug/debug-object.js";
-import { GameState } from "../models/game-state.js";
-import { ServiceLocator } from "./service-locator.js";
-import { ScreenTransitionService } from "./screen-transition-service.js";
+import { DebugObject } from "../../debug/debug-object.js";
+import { GameState } from "../../models/game-state.js";
+import { ServiceLocator } from "../service-locator.js";
+import { ScreenTransitionService } from "../screen-transition-service.js";
 import { MatchmakingService } from "./matchmaking-service.js";
-import { DebugService } from "../debug/debug-service.js";
-import { WebRTCService } from "./webrtc-service.js";
+import { DebugService } from "../../debug/debug-service.js";
+import { WebRTCService } from "../network/webrtc-service.js";
 import { TimerManagerService } from "./timer-manager-service.js";
 import { IntervalManagerService } from "./interval-manager-service.js";
-import { ServiceRegistry } from "./service-registry.js";
-import { LoadingIndicatorObject } from "../objects/common/loading-indicator-object.js";
+import { ServiceRegistry } from "../service-registry.js";
+import { LoadingIndicatorObject } from "../../objects/common/loading-indicator-object.js";
 
 export class GameLoopService {
   private context: CanvasRenderingContext2D;
