+++ src/services/gameplay/interval-manager-service.ts
