+++ src/services/gameplay/interval-service.ts
