+++ src/services/gameplay/object-orchestrator-service.ts
@@ -1,17 +1,17 @@
-import type { MultiplayerGameObject } from "../interfaces/objects/multiplayer-game-object.js";
-import { WebRTCService } from "./webrtc-service.js";
-import { GameState } from "../models/game-state.js";
-import type { WebRTCPeer } from "../interfaces/webrtc-peer.js";
-import { ObjectUtils } from "../utils/object-utils.js";
-import type { MultiplayerScreen } from "../interfaces/screens/multiplayer-screen.js";
-import { ObjectStateType } from "../enums/object-state-type.js";
-import { ScreenUtils } from "../screens/utils/screen-utils.js";
-import { WebRTCType } from "../enums/webrtc-type.js";
-import { BinaryReader } from "../utils/binary-reader-utils.js";
-import { BinaryWriter } from "../utils/binary-writer-utils.js";
-import type { ObjectType } from "../enums/object-type.js";
-import { PeerCommandHandler } from "../decorators/peer-command-handler-decorator.js";
-import { ServiceLocator } from "./service-locator.js";
+import type { MultiplayerGameObject } from "../../interfaces/objects/multiplayer-game-object.js";
+import { WebRTCService } from "../network/webrtc-service.js";
+import { GameState } from "../../models/game-state.js";
+import type { WebRTCPeer } from "../../interfaces/webrtc-peer.js";
+import { ObjectUtils } from "../../utils/object-utils.js";
+import type { MultiplayerScreen } from "../../interfaces/screens/multiplayer-screen.js";
+import { ObjectStateType } from "../../enums/object-state-type.js";
+import { ScreenUtils } from "../../screens/utils/screen-utils.js";
+import { WebRTCType } from "../../enums/webrtc-type.js";
+import { BinaryReader } from "../../utils/binary-reader-utils.js";
+import { BinaryWriter } from "../../utils/binary-writer-utils.js";
+import type { ObjectType } from "../../enums/object-type.js";
+import { PeerCommandHandler } from "../../decorators/peer-command-handler-decorator.js";
+import { ServiceLocator } from "../service-locator.js";
 
 export class ObjectOrchestratorService {
   private readonly PERIODIC_MILLISECONDS = 500;
