+++ src/services/network/matchmaking-network-service.ts
@@ -1,28 +1,28 @@
-import { TimerService } from "./timer-service.js";
-import { Match } from "../models/match.js";
-import { GamePlayer } from "../models/game-player.js";
-import { GameState } from "../models/game-state.js";
-import { MATCH_ATTRIBUTES } from "../constants/matchmaking-constants.js";
-import type { WebRTCPeer } from "../interfaces/webrtc-peer.js";
-import { EventType } from "../enums/event-type.js";
-import { LocalEvent } from "../models/local-event.js";
-import type { PlayerConnectedPayload } from "../interfaces/events/player-connected-payload.js";
-import type { PlayerDisconnectedPayload } from "../interfaces/events/player-disconnected-payload.js";
-import { WebRTCType } from "../enums/webrtc-type.js";
-import { IntervalService } from "./interval-service.js";
-import { WebSocketType } from "../enums/websocket-type.js";
-import { BinaryWriter } from "../utils/binary-writer-utils.js";
-import { BinaryReader } from "../utils/binary-reader-utils.js";
-import { PeerCommandHandler } from "../decorators/peer-command-handler-decorator.js";
-import { ServerCommandHandler } from "../decorators/server-command-handler.js";
-import { ServiceLocator } from "./service-locator.js";
+import { TimerService } from "../gameplay/timer-service.js";
+import { Match } from "../../models/match.js";
+import { GamePlayer } from "../../models/game-player.js";
+import { GameState } from "../../models/game-state.js";
+import { MATCH_ATTRIBUTES } from "../../constants/matchmaking-constants.js";
+import type { WebRTCPeer } from "../../interfaces/webrtc-peer.js";
+import { EventType } from "../../enums/event-type.js";
+import { LocalEvent } from "../../models/local-event.js";
+import type { PlayerConnectedPayload } from "../../interfaces/events/player-connected-payload.js";
+import type { PlayerDisconnectedPayload } from "../../interfaces/events/player-disconnected-payload.js";
+import { WebRTCType } from "../../enums/webrtc-type.js";
+import { IntervalService } from "../gameplay/interval-service.js";
+import { WebSocketType } from "../../enums/websocket-type.js";
+import { BinaryWriter } from "../../utils/binary-writer-utils.js";
+import { BinaryReader } from "../../utils/binary-reader-utils.js";
+import { PeerCommandHandler } from "../../decorators/peer-command-handler-decorator.js";
+import { ServerCommandHandler } from "../../decorators/server-command-handler.js";
+import { ServiceLocator } from "../service-locator.js";
 import { WebSocketService } from "./websocket-service.js";
 import { WebRTCService } from "./webrtc-service.js";
-import type { PeerConnectionListener } from "../interfaces/services/peer-connection-listener.js";
-import { EventProcessorService } from "./event-processor-service.js";
-import { TimerManagerService } from "./timer-manager-service.js";
-import { IntervalManagerService } from "./interval-manager-service.js";
-import { MatchFinderService } from "./match-finder-service.js";
+import type { PeerConnectionListener } from "../../interfaces/services/peer-connection-listener.js";
+import { EventProcessorService } from "../gameplay/event-processor-service.js";
+import { TimerManagerService } from "../gameplay/timer-manager-service.js";
+import { IntervalManagerService } from "../gameplay/interval-manager-service.js";
+import { MatchFinderService } from "../gameplay/match-finder-service.js";
 
 export class MatchmakingNetworkService implements PeerConnectionListener {
   private findMatchesTimerService: TimerService | null = null;
